requirements.txt
# Core web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# LangGraph pipeline
langgraph==0.2.50
langchain-core==0.3.20

# Groq API
groq>=0.11.0

# Data validation
pydantic==2.10.1
pydantic-settings==2.6.1

# SNMT / config files
pyyaml==6.0.2

# Batfish verification
pybatfish==2025.7.7.2423

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

# Utilities
python-dotenv==1.0.1